% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer.R
\name{confer}
\alias{confer}
\title{The Coale Model of Marital Fertility}
\usage{
confer(age)
}
\arguments{
\item{age}{vector of exact ages}
}
\description{
Computes Coale's schedule of control by age by linear interpolation
}

