% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfer.R
\name{marfer}
\alias{marfer}
\title{Coale's Marital Fertility Model}
\usage{
marfer(age, level = 1, control = 0)
}
\arguments{
\item{age}{vector of exact ages}

\item{level}{(M) scalar representing natural fertility level, default 1}

\item{control}{(m) scalar representing degree of control, default 0}
}
\description{
Computes model marital fertility rates by age
}

