% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natfer.R
\name{natfer}
\alias{natfer}
\title{Henry's Natural Fertility Schedule}
\usage{
natfer(age)
}
\arguments{
\item{age}{is a vector of exact ages}
}
\description{
Computes Henry's natural fertility schedule by linear interpolation
}

