% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnupt.R
\name{pnupt}
\alias{pnupt}
\title{The Coale-McNeil Nuptiality Model}
\usage{
pnupt(age, mean, stdev, pem = 1)
}
\arguments{
\item{age}{vector of exact ages}

\item{mean}{scalar representing mean age at marriage}

\item{stdev}{scalar representing standard deviation of age at marriage}

\item{pem}{probability of ever marrying, defaulting to 1}
}
\description{
Computes cumulative probabilities of marriage by age
}

