% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhernes.R
\name{qhernes}
\alias{qhernes}
\title{Hernes Nuptiality Model}
\usage{
qhernes(p, A, r, pem)
}
\arguments{
\item{p}{vector of cumulative probabilities to compute quantiles}

\item{A}{scalar representing attractiveness at age 15}

\item{r}{scalar representing the decay rate, so Hernes's b = exp(-r)}

\item{pem}{scalar representing the probability of ever marrying}
}
\description{
Computes quantiles of age at marriage
}

