% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnupt.R
\name{qnupt}
\alias{qnupt}
\title{The Coale-McNeil Nuptiality Model}
\usage{
qnupt(prob, mean, stdev, pem = 1)
}
\arguments{
\item{prob}{vector of probabilities of marrying}

\item{mean}{scalar representing mean age at marriage}

\item{stdev}{scalar representing standard deviation of age at marriage}

\item{pem}{probability of ever marrying, defaulting to 1}
}
\description{
Computes quantiles of age at marriage
}

